+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Improving performance of `#.opposite`.
 * Improving serialization performance.
 * Updating type declarations & adding generics.
+* Optimizing various degree methods.
 
 ## 0.17.1